requirements.txt
streamlit
google-genai
pillow
pillow-heif
python-dotenv
